requirements.txt
Django>=4.2,<5.1
wagtail>=6.2,<6.3
django-unfold
django-browser-reload
django-tailwind
djangorestframework-simplejwt
sendgrid